Lilac: ensure gene cache is initialised before tests that need it
